converted alignments to float scores
git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@27 6721810a-1124-0410-938e-bb277542c816

+++ jebl/evolution/align/NeedlemanWunsch.java
@@ -4,11 +4,11 @@ import jebl.evolution.align.scores.Scores;
 
 public class NeedlemanWunsch extends AlignSimple {
 
-    public NeedlemanWunsch(Scores sub, int d, String sq1, String sq2) {
+    public NeedlemanWunsch(Scores sub, float d, String sq1, String sq2) {
 
         super(sub, d, sq1, sq2);
         int n = this.n, m = this.m;
-        int[][] score = sub.score;
+        float[][] score = sub.score;
 
         for (int i=1; i<=n; i++) {
             F[i][0] = -d * i;
@@ -20,8 +20,8 @@ public class NeedlemanWunsch extends AlignSimple {
         }
         for (int i=1; i<=n; i++) {
             for (int j=1; j<=m; j++) {
-                int s = score[seq1.charAt(i-1)][seq2.charAt(j-1)];
-                int val = max(F[i-1][j-1]+s, F[i-1][j]-d, F[i][j-1]-d);
+                float s = score[seq1.charAt(i-1)][seq2.charAt(j-1)];
+                float val = max(F[i-1][j-1]+s, F[i-1][j]-d, F[i][j-1]-d);
                 F[i][j] = val;
                 if (val == F[i-1][j-1]+s) {
                     B[i][j] = new TracebackSimple(i-1, j-1);
